[issue-2] login분리
- login관련 기능들을 loginUtil.js로 분리
